# backend/pointsheet/static/openapi.yml
openapi: 3.1.0
info:
  title: Sim Racing Point Sheet API
  description: Sim Racing Point Sheet API
  version: 0.0.1
servers:
  - url: "http://localhost:5000"
    description: localhost
  - url: "http://pointsheet-app.com/api"
    description: local domain
security:
  - BearerAuth: []
paths:
  /auth:
    post:
      operationId: authenticateUser
      description: Authenticate a user using username and password
      tags:
        - Authentication
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthUserRequest"
      responses:
        200:
          description: Authentication Token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenReponse"
        401:
          description: Unauthorized - Invalid username or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                required:
                  - error
                  - message
              examples:
                invalidCredentials:
                  value:
                    error: "Unauthorized"
                    message: "Invalid username or password."
  /auth/register:
    post:
      operationId: registerUser
      description: Register a new user
      tags:
        - Authentication
      security: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUser"
      responses:
        204:
          description: User registered successfully
  /series:
    get:
      description: Fetch all series
      operationId: fetchAllSeries
      tags:
        - Series
      security:
        BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum:
              - started
              - closed
              - not_started
      responses:
        200:
          description: List of series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListSeries"
        401:
          description: Forbidden request
    post:
      description: Create series
      operationId: createSeries
      tags:
        - Series
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeriesRequest"
      responses:
        201:
          description: Created Series
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ResourceCreated"
        400:
          description: Bad request
  /series/{series_id}:
    get:
      description: Get series
      operationId: fetchSeriesById
      parameters:
        - in: path
          name: series_id
          required: true
          schema:
            type: string
      tags:
        - Series
      responses:
        200:
          description: Series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Series"
        404:
          description: Not found
        401:
          description: Forbidden
    delete:
      description: Delete series
      operationId: deleteSeries
      parameters:
        - in: path
          name: series_id
          schema:
            type: string
          required: true
      tags:
        - Series
      responses:
        204:
          description: Delete series
  /series/{series_id}/status:
    put:
      operationId: updateSeriesStatus
      description: Update the status of Series.
      tags:
        - Series
      parameters:
        - in: path
          name: series_id
          schema:
            type: uuid
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - started
                    - closed
                required: true
      responses:
        204:
          description: Status updated
        404:
          description: Not found
        401:
          description: Forbidden
  /series/{series_id}/events:
    post:
      description: Create event for series
      operationId: createEventForSeries
      parameters:
        - in: path
          name: series_id
          required: true
          schema:
            type: uuid
      tags:
        - Series
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvent"
      responses:
        204:
          description: Event created for series
    put:
      operationId: updateSeriesEvent
      tags:
        - Series
      parameters:
        - in: path
          name: series_id
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEvent"
      responses:
        204:
          description: Event updated
        404:
          description: Not found
  /series/{series_id}/events/{event_id}:
    delete:
      description: Remove event from series
      operationId: removeEventFromSeries
      tags:
        - Series
      parameters:
        - in: path
          name: series_id
          schema:
            type: uuid
          required: true
        - in: path
          name: event_id
          schema:
            type: uuid
          required: true
      responses:
        204:
          description: Event removed
        404:
          description: Event not found
  /events:
    get:
      operationId: getAllEvents
      description: Fetch all events not closed.
      tags:
        - Events
      responses:
        200:
          description: Event
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        401:
          description: Forbidden
    post:
      description: Create a new event
      operationId: createEvent
      tags:
        - Events
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvent"
      responses:
        201:
          description: Event successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        400:
          description: Bad request

  /events/{id}/:
    delete:
      description: Delete event
      operationId: deleteEventById
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      tags:
        - Events
      responses:
        204:
          description: Item deleted
    get:
      description: Get event
      operationId: getEventById
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      tags:
        - Events
      responses:
        200:
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"

  /events/{id}/schedule:
    post:
      description: Create schedule for event
      operationId: addSchedule
      tags:
        - Events
      parameters:
        - in: path
          name: id
          schema:
            type: string
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchedule"
      responses:
        400:
          description: Bad request
        204:
          description: Schedule created
  /events/{event_id}/schedule/{schedule_id}:
    delete:
      operationId: removeEventSchedule
      tags:
        - Events
      parameters:
        - name: event_id
          in: path
          schema:
            type: string
            format: uuid
          required: true
        - name: schedule_id
          in: path
          schema:
            type: integer
          required: true
      responses:
        404:
          description: Event not found
        204:
          description: Schedule deleted

  /events/{id}/join:
    put:
      operationId: joinEvent
      tags:
        - Events
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        204:
          description: User joined event
        400:
          description: Bad request
  /events/{id}/leave:
    put:
      operationId: leaveEvent
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      tags:
        - Events
      responses:
        204:
          description: Removed from event

  /events/{id}/schedule/{schedule_id}/results/file:
    post:
      description: Upload results file for a given event
      operationId: uploadEventResults
      tags:
        - Events
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
        - in: path
          name: schedule_id
          schema:
            type: integer
          required: true
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        201:
          description: Results file uploaded successfully
        400:
          description: Bad request
        415:
          description: Unsupported Media Type
    get:
      operationId: getResults
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: "uuid"
      tags:
        - Events
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ListOfResults"
        404:
          description: Not found
        400:
          description: Bad request
  /events/{id}/result:
    post:
      description: Upload results file for a given event
      operationId: addRaceResult
      tags:
        - Events
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRaceResult"
      responses:
        201:
          description: Results file uploaded successfully
        400:
          description: Bad request
        415:
          description: Unsupported Media Type
    get:
      operationId: getResults
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: "uuid"
      tags:
        - Events
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: "#/components/responses/ListOfResults"
        404:
          description: Not found
        400:
          description: Bad request
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  schemas:
    AuthUserRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    AuthTokenReponse:
      type: object
      properties:
        username:
          type: string
        auth_token:
          type: string
        auth_expired_in:
          type: datetime
    RegisterUser:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    Series:
      type: object
      required:
        - id
        - title
        - starts_at
        - ends_at
      properties:
        id:
          type: string
        title:
          type: string
        starts_at:
          type: datetime
        ends_at:
          type: datetime
        status:
          type: string
          enum:
            - open
            - started
            - closed
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
    ListSeries:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        starts_at:
          type: datetime
        ends_at:
          type: datetime
        status:
          type: string
          enum:
            - open
            - closed
        event_count:
          type: string
          example: 0

    CreateSeriesRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/SeriesStatus"
          example: started
        starts_at:
          type: datetime
          example: 2025-01-01
        ends_at:
          type: datetime
          example: 2025-01-01

    SeriesStatus:
      type: string
      enum:
        - started
        - closed

    Event:
      description: An event may optionally belong to a series or be standalone in the case of one-off events.
      type: object
      required:
        - id
        - title
        - starts_at
        - ends_at
        - hosts
        - status
      properties:
        id:
          type: string
        title:
          type: string
        starts_at:
          type: datetime
        ends_at:
          type: datetime
        host:
          description: The person running the event.
          type: integer
        track:
          type: string
          default: TBD
        status:
          description: Default value is open. Users can register when this is open. No new signups when in progress or closed.
          type: string
          enum:
            - open
            - closed
            - in_progress
          default: open
        series:
          type: string
          description: UUID of series
        rules:
          type: string
          description: Can contain rules like number of pit stops, tire restrictions and others for this event.
        schedule:
          description: The sequence of activities making up the event. Qualification, Practice or race.
          type: array
          items:
            $ref: "#/components/schemas/Schedule"
    CreateEvent:
      type: object
      required:
        - title
        - host
        - status
        - starts_at
        - ends_at
      properties:
        title:
          type: string
        starts_at:
          type: datetime
        ends_at:
          type: datetime
        host:
          type: string
        status:
          type: string
          enum:
            - open
            - closed
            - in_progress
        rules:
          type: string
    UpdateEvent:
      type: object
      required:
        - id
        - title
        - status
        - host
      properties:
        id:
          type: uuid
        title:
          type: string
        starts_at:
          type: datetime
        ends_at:
          type: datetime
        host:
          type: string
        status:
          type: string
          enum:
            - open
            - closed
            - in_progress
        series:
          type: string
        rules:
          type: string

    Schedule:
      type: object
      required:
        - id
        - type
        - nbr_of_laps
        - duration
      properties:
        id:
          type: int
        type:
          type: string
          enum:
            - Practice
            - Qualification
            - Race
        nbr_of_laps:
          type: integer
        duration:
          description: The duration of the schedule.
          type: string
          example: 01:45

    PointDistribution:
      description: Not used yet. TBD
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        points:
          type: object
          additionalProperties:
            type: integer
            example: { "1": 10, "2": 20 }

    CreateSchedule:
      type: object
      require:
        - type
      properties:
        type:
          type: string
          enum:
            - practice
            - qualification
            - race
        nbr_of_laps:
          type: integer
        duration:
          type: string
          example: 1h 30m
    DriverResult:
      type: object
      required:
        - driver
        - driver_id
        - position
        - best_lap
        - total
        - penalties
        - fl_points
        - total_points
        - total
      properties:
        driver_id:
          type: string
          format: "uuid"
        driver:
          type: string
          example: driver 1
        position:
          type: integer
        best_lap:
          type: string
        total:
          type: string
        penalties:
          type: integer
        fl_points:
          type: integer
        points:
          type: integer
        total_points:
          type: integer
    RaceResult:
      type: object
      properties:
        schedule_id:
          type: string
        mark_down:
          type: string
        upload_file:
          type: string
        result:
          type: array
          items:
            $ref: "#/components/schemas/DriverResult"
    CreateRaceResult:
      type: object
      properties:
        schedule_id:
          type: string
        result:
          type: array
          items:
            $ref: "#/components/schemas/DriverResult"
  responses:
    ListOfResults:
      type: array
      items:
        $ref: "#/components/schemas/RaceResult"
    ResourceCreated:
      type: object
      properties:
        resource:
          type: string
          example: uuid or id